Apply @YarShev's suggestions
