+++ modin/utils.py
@@ -18,12 +18,10 @@ import types
 from typing import (
     Any,
     Callable,
-    Iterator,
     List,
     Mapping,
     Optional,
     Union,
-    Tuple,
     TypeVar,
 )
 import re
@@ -44,14 +42,7 @@ import numpy as np
 
 from pandas.util._decorators import Appender
 from pandas.util._print_versions import _get_sys_info, _get_dependency_info  # type: ignore[attr-defined]
-from pandas._typing import (  # type: ignore[attr-defined]
-    AggFuncType,
-    AggFuncTypeBase,
-    AggFuncTypeDict,
-    IndexLabel,
-    JSONSerializable,
-)
-from pandas.core.dtypes.common import is_list_like
+from pandas._typing import JSONSerializable
 
 from modin.config import Engine, StorageFormat, IsExperimental
 from modin._version import get_versions
@@ -749,50 +740,3 @@ def show_versions(as_json: Union[str, bool] = False) -> None:
             print(f"\n{name} dependencies\n{'-' * (len(name) + 13)}")
             for k, v in d.items():
                 print(f"{k:<{maxlen}}: {v}")
-
-
-def walk_aggregation_dict(
-    agg_dict: AggFuncTypeDict,
-) -> Iterator[Tuple[IndexLabel, AggFuncTypeBase, Optional[str], bool]]:
-    """
-    Walk over an aggregation dictionary.
-
-    Parameters
-    ----------
-    agg_dict : AggFuncTypeDict
-
-    Yields
-    ------
-    (col: IndexLabel, func: AggFuncTypeBase, func_name: Optional[str], col_renaming_required: bool)
-        Yield an aggregation function with its metadata:
-            - `col`: column name to apply the function.
-            - `func`: aggregation function to apply to the column.
-            - `func_name`: custom function name that was specified in the dict.
-            - `col_renaming_required`: whether it's required to rename the
-                `col` into ``(col, func_name)``.
-    """
-
-    def walk(
-        key: IndexLabel, value: AggFuncType, depth: int = 0
-    ) -> Iterator[Tuple[IndexLabel, AggFuncTypeBase, Optional[str], bool]]:
-        col_renaming_required = bool(depth)
-
-        if isinstance(value, (list, tuple)):
-            if depth == 0:
-                for val in value:
-                    yield from walk(key, val, depth + 1)
-            elif depth == 1:
-                if len(value) != 2:
-                    raise ValueError(
-                        f"Incorrect rename format. Renamer must consist of exactly two elements, got {len(value)=}."
-                    )
-                func_name, func = value
-                yield key, func, func_name, col_renaming_required
-            else:
-                # pandas doesn't support this as well
-                raise NotImplementedError("Nested renaming is not supported.")
-        else:
-            yield key, value, None, col_renaming_required
-
-    for key, value in agg_dict.items():
-        yield from walk(key, value)
